+++ src/Farmer/Builders/Builders.Functions.fs
@@ -384,6 +384,8 @@ type FunctionsConfig = {
                 DisableBlobPublicAccess = None
                 DisableSharedKeyAccess = None
                 DefaultToOAuthAuthentication = None
+                ImmutableStorageWithVersioning = None
+                RequireInfrastructureEncryption = None
               }
             | _ -> ()