+++ src/Farmer/Builders/Builders.Vm.fs
@@ -264,6 +264,8 @@ type VmConfig = {
                     DisableBlobPublicAccess = None
                     DisableSharedKeyAccess = None
                     DefaultToOAuthAuthentication = None
+                    ImmutableStorageWithVersioning = None
+                    RequireInfrastructureEncryption = None
                   }
                 | Some _
                 | None -> ()